% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyhifi_obj.R
\name{get_non_absrb}
\alias{get_non_absrb}
\title{Get non-absorbing states}
\usage{
get_non_absrb(M.phf)
}
\arguments{
\item{M.phf}{matrix}
}
\value{
vector of names.
}
\description{
Get non-absorbing states
}
