% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ray_disc.R
\name{greedy_rayDISC}
\alias{greedy_rayDISC}
\title{Mod Ray Disc}
\usage{
greedy_rayDISC(phy, data, ntraits = 1, charnum = 1, rate.mat = NULL,
  model = c("ARD"), node.states = c("marginal"),
  state.recon = c("subsequently"), lewis.asc.bias = FALSE, p = NULL,
  root.p = NULL, ip = NULL, lb = 0, ub = 100, verbose = TRUE,
  diagn = FALSE, do.ans = FALSE, do.thorough = FALSE)
}
\arguments{
\item{M}{matrix}
}
\value{
vector of names.
}
\description{
Mod Ray Disc
}
\examples{
phyhifi()
}
