% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{make_init_matrix}
\alias{make_init_matrix}
\title{Create intial matix for run_hifi}
\usage{
make_init_matrix(part_scheme, set.NA = NULL)
}
\arguments{
\item{part_scheme}{partitioning scheme that is a vector where element id is an observable state
while the element value is the number of hidden changes}

\item{set.NA}{a vector that sets main diagonal elements (i.k, i.l) to NA. Good to use if some observable state
is known to lack any hidden states. For example if observable state 1.x lacks hidden states than
specify set.NA=c(1)}
}
\value{
Matrix
}
\description{
Creates intial matrix for evolutinary agorithm. NA's mean constant (0 or diagonal)
cells, 0's mean cells that can potentially change, 1's mean cells with the rate > 0.
}
\examples{
#part_scheme<-c(5, 2, 3)
#names(part_scheme)<-c("s1", "s2", "s3")
#make_init_matrix(part_scheme, set.NA=c(2,3) )
#make_init_matrix(part_scheme)
}
