% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyhifi_obj.R
\name{phyhifi}
\alias{phyhifi}
\title{Create PhyHiFi object}
\usage{
phyhifi(M = NULL, rate = NULL, Ln = NULL, AIC = NULL,
  AICc = NULL, elapsed.time = NULL, Cor.sol = FALSE, count = 1)
}
\arguments{
\item{M}{matrix}
}
\value{
vector of names.
}
\description{
Create PhyHiFi object
}
\examples{
phyhifi()
}
