% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyhifi_obj.R
\name{unique_matrix_phf2}
\alias{unique_matrix_phf2}
\title{select unique matrices given two sets: check M.phf.list2 against M.phf.list1}
\usage{
unique_matrix_phf2(M.phf.list1, M.phf.list2, count = FALSE)
}
\description{
select unique matrices given two sets: check M.phf.list2 against M.phf.list1
}
\examples{
phyhifi()
}
